+++ modules/tinymce/src/core/test/ts/browser/caret/FakeCaretTest.ts
@@ -1,9 +1,9 @@
 import { after, before, describe, it } from '@ephox/bedrock-client';
 import { Fun } from '@ephox/katamari';
-import { SugarElement } from '@ephox/sugar';
+import { Attribute, Html, SelectorFilter, SelectorFind, SugarElement, SugarNode, Traverse } from '@ephox/sugar';
 import { assert } from 'chai';
 
-import $ from 'tinymce/core/api/dom/DomQuery';
+import DOMUtils from 'tinymce/core/api/dom/DOMUtils';
 import { FakeCaret, isFakeCaretTarget } from 'tinymce/core/caret/FakeCaret';
 import { isFakeCaretTableBrowser } from 'tinymce/core/keyboard/TableNavigation';
 import * as Zwsp from 'tinymce/core/text/Zwsp';
@@ -15,13 +15,14 @@ describe('browser.tinymce.core.caret.FakeCaretTest', () => {
   const viewBlock = ViewBlock.bddSetup();
   let fakeCaret: FakeCaret;
 
-  const getRoot = viewBlock.get;
+  const getRoot = () => SugarElement.fromDom(viewBlock.get());
 
   before(() => {
     const mockEditor: any = {
-      getParam: Fun.constant('p')
+      getParam: Fun.constant('p'),
+      dom: DOMUtils(document)
     };
-    fakeCaret = FakeCaret(mockEditor, getRoot(), isBlock, Fun.always);
+    fakeCaret = FakeCaret(mockEditor, getRoot().dom, isBlock, Fun.always);
   });
 
   after(() => {
@@ -33,59 +34,59 @@ describe('browser.tinymce.core.caret.FakeCaretTest', () => {
   };
 
   it('show/hide (before, block)', () => {
-    $(getRoot()).html('<div>a</div>');
+    Html.set(getRoot(), '<div>a</div>');
 
-    const rng = fakeCaret.show(true, $<HTMLDivElement>('div', getRoot())[0]);
-    const $fakeCaretElm = $(getRoot()).children();
+    const rng = fakeCaret.show(true, SelectorFind.descendant(getRoot(), 'div').getOrDie().dom);
+    const fakeCaretElm = Traverse.children(getRoot())[0] as SugarElement<HTMLElement>;
 
-    assert.equal($fakeCaretElm[0].nodeName, 'P');
-    assert.equal($fakeCaretElm.attr('data-mce-caret'), 'before');
-    CaretAsserts.assertRange(rng, CaretAsserts.createRange($fakeCaretElm[0], 0, $fakeCaretElm[0], 0));
+    assert.equal(SugarNode.name(fakeCaretElm), 'p');
+    assert.equal(Attribute.get(fakeCaretElm, 'data-mce-caret'), 'before');
+    CaretAsserts.assertRange(rng, CaretAsserts.createRange(fakeCaretElm.dom, 0, fakeCaretElm.dom, 0));
 
     fakeCaret.hide();
-    assert.lengthOf($('*[data-mce-caret]', getRoot()), 0);
+    assert.lengthOf(SelectorFilter.descendants(getRoot(), '*[data-mce-caret]'), 0);
   });
 
   it('show/hide (after, block)', () => {
-    $(getRoot()).html('<div>a</div>');
+    Html.set(getRoot(), '<div>a</div>');
 
-    const rng = fakeCaret.show(false, $<HTMLDivElement>('div', getRoot())[0]);
-    const $fakeCaretElm = $(getRoot()).children();
+    const rng = fakeCaret.show(false, SelectorFind.descendant(getRoot(), 'div').getOrDie().dom);
+    const fakeCaretElm = Traverse.children(getRoot())[1] as SugarElement<HTMLElement>;
 
-    assert.equal($fakeCaretElm[1].nodeName, 'P');
-    assert.equal($fakeCaretElm.eq(1).attr('data-mce-caret'), 'after');
-    CaretAsserts.assertRange(rng, CaretAsserts.createRange($fakeCaretElm[1], 0, $fakeCaretElm[1], 0));
+    assert.equal(SugarNode.name(fakeCaretElm), 'p');
+    assert.equal(Attribute.get(fakeCaretElm, 'data-mce-caret'), 'after');
+    CaretAsserts.assertRange(rng, CaretAsserts.createRange(fakeCaretElm.dom, 0, fakeCaretElm.dom, 0));
 
     fakeCaret.hide();
-    assert.lengthOf($('*[data-mce-caret]', getRoot()), 0);
+    assert.lengthOf(SelectorFilter.descendants(getRoot(), '*[data-mce-caret]'), 0);
   });
 
   it('show/hide (before, inline)', () => {
-    $(getRoot()).html('<span>a</span>');
+    Html.set(getRoot(), '<span>a</span>');
 
-    const rng = fakeCaret.show(true, $<HTMLSpanElement>('span', getRoot())[0]);
-    const $fakeCaretText = $(getRoot()).contents();
+    const rng = fakeCaret.show(true, SelectorFind.descendant(getRoot(), 'span').getOrDie().dom);
+    const fakeCaretText = Traverse.children(getRoot());
 
-    assert.equal($fakeCaretText[0].nodeName, '#text');
-    assert.equal(($fakeCaretText[0] as Text).data, Zwsp.ZWSP);
-    CaretAsserts.assertRange(rng, CaretAsserts.createRange($fakeCaretText[0], 1));
+    assert.equal(SugarNode.name(fakeCaretText[0]), '#text');
+    assert.equal((fakeCaretText[0].dom as Text).data, Zwsp.ZWSP);
+    CaretAsserts.assertRange(rng, CaretAsserts.createRange(fakeCaretText[0].dom, 1));
 
     fakeCaret.hide();
-    assert.equal($(getRoot()).contents()[0].nodeName, 'SPAN');
+    assert.equal(Traverse.children(getRoot())[0].dom.nodeName, 'SPAN');
   });
 
   it('show/hide (after, inline)', () => {
-    $(getRoot()).html('<span>a</span>');
+    Html.set(getRoot(), '<span>a</span>');
 
-    const rng = fakeCaret.show(false, $<HTMLSpanElement>('span', getRoot())[0]);
-    const $fakeCaretText = $(getRoot()).contents();
+    const rng = fakeCaret.show(false, SelectorFind.descendant(getRoot(), 'span').getOrDie().dom);
+    const fakeCaretText = Traverse.children(getRoot());
 
-    assert.equal($fakeCaretText[1].nodeName, '#text');
-    assert.equal(($fakeCaretText[1] as Text).data, Zwsp.ZWSP);
-    CaretAsserts.assertRange(rng, CaretAsserts.createRange($fakeCaretText[1], 1));
+    assert.equal(SugarNode.name(fakeCaretText[1]), '#text');
+    assert.equal((fakeCaretText[1].dom as Text).data, Zwsp.ZWSP);
+    CaretAsserts.assertRange(rng, CaretAsserts.createRange(fakeCaretText[1].dom, 1));
 
     fakeCaret.hide();
-    assert.equal($(getRoot()).contents()[0].nodeName, 'SPAN');
+    assert.equal(Traverse.children(getRoot())[0].dom.nodeName, 'SPAN');
   });
 
   it('getCss', () => {
@@ -93,14 +94,14 @@ describe('browser.tinymce.core.caret.FakeCaretTest', () => {
   });
 
   it('show before TD', () => {
-    getRoot().innerHTML = '<table><tr><td contenteditable="false">x</td></tr></table>';
-    const rng = fakeCaret.show(false, $<HTMLTableCellElement>('td', getRoot())[0]);
+    getRoot().dom.innerHTML = '<table><tr><td contenteditable="false">x</td></tr></table>';
+    const rng = fakeCaret.show(false, SelectorFind.descendant(getRoot(), 'td').getOrDie().dom);
     assert.isNull(rng, 'Should be null since TD is not a valid caret target');
   });
 
   it('show before TH', () => {
-    getRoot().innerHTML = '<table><tr><th contenteditable="false">x</th></tr></table>';
-    const rng = fakeCaret.show(false, $<HTMLTableCellElement>('th', getRoot())[0]);
+    getRoot().dom.innerHTML = '<table><tr><th contenteditable="false">x</th></tr></table>';
+    const rng = fakeCaret.show(false, SelectorFind.descendant(getRoot(), 'th').getOrDie().dom);
     assert.isNull(rng, 'Should be null since TH is not a valid caret target');
   });