+++ modules/tinymce/src/plugins/lists/test/ts/browser/BackspaceDeleteInlineTest.ts
@@ -3,7 +3,6 @@ import { Html, Insert, Remove, SugarBody, SugarElement } from '@ephox/sugar';
 import { TinyHooks } from '@ephox/wrap-mcagar';
 import { assert } from 'chai';
 
-import DomQuery from 'tinymce/core/api/dom/DomQuery';
 import Editor from 'tinymce/core/api/Editor';
 import Plugin from 'tinymce/plugins/lists/Plugin';
 import Theme from 'tinymce/themes/silver/Theme';
@@ -46,8 +45,8 @@ describe('browser.tinymce.plugins.lists.BackspaceDeleteInlineTest', () => {
     const body = editor.getBody();
     editor.selection.setCursorLocation(body.firstChild.firstChild, 0);
     editor.plugins.lists.backspaceDelete();
-    assert.lengthOf(DomQuery('#lists ul'), 3);
-    assert.lengthOf(DomQuery('#lists li'), 3);
+    assert.lengthOf(editor.dom.select('#lists ul'), 3);
+    assert.lengthOf(editor.dom.select('#lists li'), 3);
   });
 
   it('TBA: Delete at end of LI on body UL', () => {
@@ -55,7 +54,7 @@ describe('browser.tinymce.plugins.lists.BackspaceDeleteInlineTest', () => {
     const body = editor.getBody();
     editor.selection.setCursorLocation(body.firstChild.firstChild, 1);
     editor.plugins.lists.backspaceDelete(true);
-    assert.lengthOf(DomQuery('#lists ul'), 3);
-    assert.lengthOf(DomQuery('#lists li'), 3);
+    assert.lengthOf(editor.dom.select('#lists ul'), 3);
+    assert.lengthOf(editor.dom.select('#lists li'), 3);
   });
 });